requirements.txt
discord.py
django
numpy
PILLOW
pycountry
pyUFbr
requests